Add files via upload
I took the manual approach the author described and tweak it to work for our project. For the code, he has on intensifying promotable pieces (https://github.com/joeymeyer/raspberryturk/blob/master/notebooks/chess_piece_classification.ipynb) was there anything that needed to be done for us to work with it?