Update model_evaluation.py
replaces repetitive code with function make_move(). There is a bit more I could replace with a function but I found it looked gross bc I had to replace the break statements with a boolean variable and had to include conditionals to check if the game was not over after every turn.